+++ internal/checks/alerts_annotation_test.go
@@ -14,7 +14,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "ignores recording rules",
 			content:     "- record: foo\n  expr: sum(foo) without(\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -23,7 +23,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "doesn't ignore rules with syntax errors",
 			content:     "- alert: foo\n  expr: sum(foo) without(\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -44,7 +44,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "no annotations / required",
 			content:     "- alert: foo\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -65,7 +65,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "no annotations / not required",
 			content:     "- alert: foo\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, false, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, false, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -74,7 +74,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "missing annotation / required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    foo: bar\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -95,7 +95,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "missing annotation / not required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    foo: bar\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, false, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, false, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -104,7 +104,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "wrong annotation value / required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    severity: bar\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -125,7 +125,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "wrong annotation value / not required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    severity: bar\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, false, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical"), nil, false, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -146,7 +146,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "valid annotation / required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    severity: info\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -155,7 +155,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "valid annotation / not required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    severity: info\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, false, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("severity"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, false, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -164,7 +164,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "templated annotation value / passing",
 			content:     "- alert: foo\n  expr: sum(foo)\n  for: 5m\n  annotations:\n    for: 5m\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("for"), nil, checks.MustTemplatedRegexp("{{ $for }}"), nil, true, checks.Bug)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("for"), nil, checks.MustTemplatedRegexp("{{ $for }}"), nil, true, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -173,7 +173,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "templated annotation value / passing",
 			content:     "- alert: foo\n  expr: sum(foo)\n  for: 5m\n  annotations:\n    for: 4m\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("for"), nil, checks.MustTemplatedRegexp("{{ $for }}"), nil, true, checks.Bug)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("for"), nil, checks.MustTemplatedRegexp("{{ $for }}"), nil, true, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -194,7 +194,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "valid annotation key regex / required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    annotation_1: info\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -203,7 +203,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "valid annotation key regex / not required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    annotation_1: info\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, false, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical|info|debug"), nil, false, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -212,7 +212,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "wrong annotation key regex value / required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    annotation_1: bar\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical"), nil, true, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical"), nil, true, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -233,7 +233,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "wrong annotation key regex value / not required",
 			content:     "- alert: foo\n  expr: sum(foo)\n  annotations:\n    annotation_1: bar\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical"), nil, false, checks.Warning)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("annotation_.*"), nil, checks.MustTemplatedRegexp("critical"), nil, false, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -254,7 +254,7 @@ func TestAnnotationCheck(t *testing.T) {
 			description: "invalid value / token / valueRe",
 			content:     "- alert: foo\n  expr: rate(foo[1m])\n  annotations:\n    components: api db\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("components"), checks.MustRawTemplatedRegexp("\\w+"), checks.MustTemplatedRegexp("api|memcached"), nil, false, checks.Bug)
+				return checks.NewAnnotationCheck(checks.MustTemplatedRegexp("components"), checks.MustRawTemplatedRegexp("\\w+"), checks.MustTemplatedRegexp("api|memcached"), nil, false, "rule comment", checks.Bug)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -266,6 +266,7 @@ func TestAnnotationCheck(t *testing.T) {
 						},
 						Reporter: checks.AnnotationCheckName,
 						Text:     "`components` annotation value `db` must match `^api|memcached$`.",
+						Details:  "Rule comment: rule comment",
 						Severity: checks.Bug,
 					},
 				}
@@ -281,6 +282,7 @@ func TestAnnotationCheck(t *testing.T) {
 					nil,
 					[]string{"api", "memcached", "storage", "prometheus", "kvm", "mysql", "memsql", "haproxy", "postgresql"},
 					false,
+					"rule comment",
 					checks.Bug,
 				)
 			},
@@ -294,7 +296,7 @@ func TestAnnotationCheck(t *testing.T) {
 						},
 						Reporter: checks.AnnotationCheckName,
 						Text:     "`components` annotation value `db` is not one of valid values.",
-						Details:  "List of allowed values:\n\n- `api`\n- `memcached`\n- `storage`\n- `prometheus`\n- `kvm`\n- `mysql`\n\nAnd 3 other value(s).",
+						Details:  "List of allowed values:\n\n- `api`\n- `memcached`\n- `storage`\n- `prometheus`\n- `kvm`\n- `mysql`\n\nAnd 3 other value(s).\nRule comment: rule comment",
 						Severity: checks.Bug,
 					},
 				}