+++ internal/checks/query_cost_test.go
@@ -44,7 +44,7 @@ func TestCostCheck(t *testing.T) {
 			description: "ignores rules with syntax errors",
 			content:     "- record: foo\n  expr: sum(foo) without(\n",
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems:   noProblems,
@@ -53,7 +53,7 @@ func TestCostCheck(t *testing.T) {
 			description: "empty response",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -83,7 +83,7 @@ func TestCostCheck(t *testing.T) {
 			description: "response timeout",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: func(uri string) *promapi.FailoverGroup {
 				return simpleProm("prom", uri, time.Millisecond*50, true)
@@ -115,7 +115,7 @@ func TestCostCheck(t *testing.T) {
 			description: "bad request",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -145,7 +145,7 @@ func TestCostCheck(t *testing.T) {
 			description: "connection refused",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: func(_ string) *promapi.FailoverGroup {
 				return simpleProm("prom", "http://127.0.0.1:1111", time.Second*5, false)
@@ -168,7 +168,7 @@ func TestCostCheck(t *testing.T) {
 			description: "1 result",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -209,7 +209,7 @@ func TestCostCheck(t *testing.T) {
 			description: "7 results",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -260,7 +260,7 @@ func TestCostCheck(t *testing.T) {
 			description: "7 result with MB",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -311,7 +311,7 @@ func TestCostCheck(t *testing.T) {
 			description: "7 results with 1 series max (1KB bps)",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 1, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 1, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -362,7 +362,7 @@ func TestCostCheck(t *testing.T) {
 			description: "6 results with 5 series max",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 5, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 5, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -408,7 +408,7 @@ func TestCostCheck(t *testing.T) {
 			description: "7 results with 5 series max / infi",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 5, 0, 0, 0, checks.Information)
+				return checks.NewCostCheck(prom, 5, 0, 0, 0, "rule comment", checks.Information)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -420,6 +420,7 @@ func TestCostCheck(t *testing.T) {
 						},
 						Reporter: "query/cost",
 						Text:     costText("prom", uri, 7) + maxSeriesText(5) + ".",
+						Details:  "Rule comment: rule comment",
 						Severity: checks.Information,
 					},
 				}
@@ -458,7 +459,7 @@ func TestCostCheck(t *testing.T) {
   expr: 'sum({__name__="foo"})'
 `,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, 0, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, 0, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -509,7 +510,7 @@ func TestCostCheck(t *testing.T) {
 			description: "1s eval, 5s limit",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 0, 0, time.Second*5, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 0, 0, time.Second*5, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -557,7 +558,7 @@ func TestCostCheck(t *testing.T) {
 			description: "stats",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 100, 10, time.Second*5, checks.Bug)
+				return checks.NewCostCheck(prom, 0, 100, 10, time.Second*5, "", checks.Bug)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -636,7 +637,7 @@ func TestCostCheck(t *testing.T) {
 			description: "stats - info",
 			content:     content,
 			checker: func(prom *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewCostCheck(prom, 0, 100, 10, time.Second*5, checks.Information)
+				return checks.NewCostCheck(prom, 0, 100, 10, time.Second*5, "some text", checks.Information)
 			},
 			prometheus: newSimpleProm,
 			problems: func(uri string) []checks.Problem {
@@ -648,6 +649,7 @@ func TestCostCheck(t *testing.T) {
 						},
 						Reporter: "query/cost",
 						Text:     costText("prom", uri, 1) + memUsageText("4.0KiB") + ".",
+						Details:  "Rule comment: some text",
 						Severity: checks.Information,
 					},
 					{
@@ -657,6 +659,7 @@ func TestCostCheck(t *testing.T) {
 						},
 						Reporter: "query/cost",
 						Text:     totalSamplesText("prom", uri, 200, 100),
+						Details:  "Rule comment: some text",
 						Severity: checks.Information,
 					},
 					{
@@ -666,6 +669,7 @@ func TestCostCheck(t *testing.T) {
 						},
 						Reporter: "query/cost",
 						Text:     peakSamplesText("prom", uri, 20, 10),
+						Details:  "Rule comment: some text",
 						Severity: checks.Information,
 					},
 					{
@@ -675,6 +679,7 @@ func TestCostCheck(t *testing.T) {
 						},
 						Reporter: "query/cost",
 						Text:     evalDurText("prom", uri, "5s100ms", "5s"),
+						Details:  "Rule comment: some text",
 						Severity: checks.Information,
 					},
 				}