+++ internal/checks/rule_for_test.go
@@ -24,7 +24,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "recording rule",
 			content:     "- record: foo\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, 0, 0, checks.Bug)
+				return checks.NewRuleForCheck(checks.RuleForFor, 0, 0, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -33,7 +33,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, no for, 0-0",
 			content:     "- alert: foo\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, 0, 0, checks.Bug)
+				return checks.NewRuleForCheck(checks.RuleForFor, 0, 0, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -42,7 +42,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:1m, 0-0",
 			content:     "- alert: foo\n  for: 1m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, 0, 0, checks.Bug)
+				return checks.NewRuleForCheck(checks.RuleForFor, 0, 0, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -51,7 +51,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:1m, 1s-0",
 			content:     "- alert: foo\n  for: 1m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, time.Second, 0, checks.Bug)
+				return checks.NewRuleForCheck(checks.RuleForFor, time.Second, 0, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -60,7 +60,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:1m, 1s-2m",
 			content:     "- alert: foo\n  for: 1m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, time.Second, time.Minute*2, checks.Bug)
+				return checks.NewRuleForCheck(checks.RuleForFor, time.Second, time.Minute*2, "", checks.Bug)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -69,7 +69,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:4m, 5m-10m",
 			content:     "- alert: foo\n  for: 4m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, time.Minute*5, time.Minute*10, checks.Warning)
+				return checks.NewRuleForCheck(checks.RuleForFor, time.Minute*5, time.Minute*10, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -90,7 +90,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:5m, 1s-2m",
 			content:     "- alert: foo\n  for: 5m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, time.Second, time.Minute*2, checks.Warning)
+				return checks.NewRuleForCheck(checks.RuleForFor, time.Second, time.Minute*2, "some text", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {
@@ -102,6 +102,7 @@ func TestRuleForCheck(t *testing.T) {
 						},
 						Reporter: "rule/for",
 						Text:     forMax("for", "2m"),
+						Details:  "Rule comment: some text",
 						Severity: checks.Warning,
 					},
 				}
@@ -111,7 +112,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:1d, 5m-0",
 			content:     "- alert: foo\n  for: 1d\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForFor, time.Minute*5, 0, checks.Warning)
+				return checks.NewRuleForCheck(checks.RuleForFor, time.Minute*5, 0, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -120,7 +121,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, for:14m, 5m-10m, keep_firing_for enforced",
 			content:     "- alert: foo\n  for: 14m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForKeepFiringFor, 0, time.Minute*10, checks.Warning)
+				return checks.NewRuleForCheck(checks.RuleForKeepFiringFor, 0, time.Minute*10, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems:   noProblems,
@@ -129,7 +130,7 @@ func TestRuleForCheck(t *testing.T) {
 			description: "alerting rule, keep_firing_for:4m, 5m-10m",
 			content:     "- alert: foo\n  keep_firing_for: 4m\n  expr: sum(foo)\n",
 			checker: func(_ *promapi.FailoverGroup) checks.RuleChecker {
-				return checks.NewRuleForCheck(checks.RuleForKeepFiringFor, time.Minute*5, time.Minute*10, checks.Warning)
+				return checks.NewRuleForCheck(checks.RuleForKeepFiringFor, time.Minute*5, time.Minute*10, "", checks.Warning)
 			},
 			prometheus: noProm,
 			problems: func(_ string) []checks.Problem {